boot: Implement multiboot1 to enable vbe mode
I had to imblement multiboot to tell the bootloader to enable vbe mode.
I chose multiboot1 because it was easier to implement and included all
I needed. Multiboot2 is needlessly complex.
Moving forwards I will also change the style of the project to use
snake_case for everything and globals in all caps.