Makefile: Fix no recompile bug if edited standalone header
The way the build was setup changes to header files were not taken into
account. So you could effectively change code like a variable
in a header and make wouldn't recompile your project.

This is frustrating so I changed it so that a change to a standalone
header recompiles the whole project. This is not efficient but I don't
see a way to do this without the complexity of the make file exploding.

Changes to a normal header will recompile only the relevant translation unit.
The catch for this is that every .c file must have a corresponding
header otherwise make will not recompile the translation unit because
the prerequisite header in the recipe doesn't exist.